Git webhook (#504)
* api server support building irectly from git repository

* api server support building irectly from git repository

* algorithm-builder - add support for clonnig repo by commit id

* algorithm-builder - add support for clonnig repo by commit id

* fix PR issues

* [api-server] - add branch support

+++ core/api-server/lib/service/builds.js
@@ -82,10 +82,20 @@ class Builds {
             buildId = this._createBuildID(algorithm.name);
             const putStream = await storageManager.hkubeBuilds.putStream({ buildId, data: fse.createReadStream(file.path) });
             merge(algorithm, { version, fileInfo: { path: putStream.path } });
-            const { env, name, fileInfo } = algorithm;
-            await this.startBuild({ buildId, algorithmName: name, env, version, fileExt: fileInfo.fileExt });
+            const { env, name, fileInfo, type } = algorithm;
+            await this.startBuild({ buildId, algorithmName: name, env, version, fileExt: fileInfo.fileExt, type });
         }
         return { algorithm, buildId, messages };
+    }
+
+    async createBuildFromGitRepository(payload) {
+        const messages = [];
+        const version = payload.gitRepository.commit.id;
+        const buildId = this._createBuildID(payload.name);
+        const { env, name, gitRepository, entryPoint, type } = payload;
+        validator.validateAlgorithmBuildFromGit({ env });
+        await this.startBuild({ buildId, version, env, algorithmName: name, gitRepository, entryPoint, type });
+        return { payload, buildId, messages };
     }
 
     async _newAlgorithm(file, oldAlgorithm, newAlgorithm) {
@@ -96,7 +106,7 @@ class Builds {
     }
 
     async removeFile(file) {
-        if (file.path) {
+        if (file && file.path) {
             await fse.remove(file.path);
         }
     }
